Put all wrapper functions in networkx.addons.metis.__init__.py
